refactor(controller): split HomeController god-object into Demo + Profile specialists 🔨
HomeController trimmed from 14 actions to 8 focused home/error page actions.
DemoController forged: RawSqlSearch, ViewLogicCalculator, CallbackHellWeather, InlineCssHell.
ProfileController forged: GodObjectProfile, GodObjectProfileUpdate.
Both new controllers use [Route("Home/[action]")] — all existing URLs preserved.
Views relocated to Views/Demo/ and Views/Profile/ to match controller names.
Layout nav updated to reference new controller names for correct URL generation.
Tests split: DemoControllerTests + ProfileControllerTests replace old HomeController anti-pattern/CSRF tests.
HomeControllerTests updated to reflect single-param constructor.
architecture.md updated with new controller table and view folder layout.

Part of #122
