+++ Views/Shared/_Layout.cshtml
@@ -91,15 +91,15 @@
                             </a>
                             <ul class="dropdown-menu" aria-labelledby="antiPatternDropdown"
                                 style="background: linear-gradient(135deg, #1a1a2e, #16213e);">
-                                <li><a class="dropdown-item" asp-controller="Home" asp-action="GodObjectProfile"
+                                <li><a class="dropdown-item" asp-controller="Profile" asp-action="GodObjectProfile"
                                         style="color: #ff6b6b;">🗃️ God Object Profile</a></li>
-                                <li><a class="dropdown-item" asp-controller="Home" asp-action="RawSqlSearch"
+                                <li><a class="dropdown-item" asp-controller="Demo" asp-action="RawSqlSearch"
                                         style="color: #4ecdc4;">💉 Raw SQL Search</a></li>
-                                <li><a class="dropdown-item" asp-controller="Home" asp-action="ViewLogicCalculator"
+                                <li><a class="dropdown-item" asp-controller="Demo" asp-action="ViewLogicCalculator"
                                         style="color: #667eea;">🧮 View Logic Calculator</a></li>
-                                <li><a class="dropdown-item" asp-controller="Home" asp-action="CallbackHellWeather"
+                                <li><a class="dropdown-item" asp-controller="Demo" asp-action="CallbackHellWeather"
                                         style="color: #f093fb;">🌊 Callback Hell Weather</a></li>
-                                <li><a class="dropdown-item" asp-controller="Home" asp-action="InlineCssHell"
+                                <li><a class="dropdown-item" asp-controller="Demo" asp-action="InlineCssHell"
                                         style="color: #ffd700;">🎨 Inline CSS Hell</a></li>
                                 <li><a class="dropdown-item" asp-controller="Achievement" asp-action="SynchronousAntiPattern"
                                         style="color: #ff9f43;">🏀 Sync Achievement Checker</a></li>
